Rename authentication client tests file
